// Alhamla/assets/types/product.type.ts
export type Product = {
    blocked: boolean;
    category_name_ar: string;
    category_name_en: string;
    company_actvie: boolean;
    company_name_ar: string;
    company_name_en: string;
    company_photo: string;
    id_cat: string;
    id_com: string;
    id_pro: string;
    id_state: string;
    product_active: boolean;
    product_description: string;
    product_name_ar: string;
    product_name_en: string;
    product_photo: string;
    rateproduct: string;
    state_name_ar: string;
    state_name_en: string;
  };
  